Project XI

+++ src/utils/api.js
@@ -68,24 +68,24 @@ class Api{
         })
     }
 
-    setCardLike(cardId){
-        return fetch(`${this._urlRequest}/cards/${cardId}/likes`, {
-            method: 'PUT',
-            headers: this._headers
-            })
-            .then(res => {
-                return this._checkResponse(res)
-            })
-    }
-
-    setRemoveLike(cardId){
-        return fetch(`${this._urlRequest}/cards/${cardId}/likes`, {
-            method: 'DELETE',
-            headers: this._headers
-            })
-            .then(res => {
-                return this._checkResponse(res)
-            })
+    changeLikeStatus(cardId, isLiked){
+        if (!isLiked){
+            return fetch(`${this._urlRequest}/cards/${cardId}/likes`, {
+                method: 'PUT',
+                headers: this._headers
+                })
+                .then(res => {
+                    return this._checkResponse(res)
+                })
+        } else {
+            return fetch(`${this._urlRequest}/cards/${cardId}/likes`, {
+                method: 'DELETE',
+                headers: this._headers
+                })
+                .then(res => {
+                    return this._checkResponse(res)
+                })
+        }
     }
     
 }
